Call GetNCVersionStatus API to validate the datapath setup (#686)
This change calls GetNCVersionStatus API for NMAgent to validate if the data path is set up. If the data path is set up, the NC version with CNS matches with the NC version programmed by NMAgent. A conservative approach is taken to check if the data path is set up i.e. only if GetNCVersionStatus successfully returns NC version mismatch, it is treated as error. All the other cases including error in calling this API are considered success.
This is multitenancy only change.