NPM adhering to both ingress and egress rules (#765)
This change will help evaluate both INGRESS and EGRESS rules before accepting/taking a decision on a packet. NPM will now MARK a packet for ingress/egress and RETURN the MARK'ed packet. Then packet will be accepted in the main chain after all the ingress and egress rules are processed.

* first pass trying to return instead of accept

* Adding initial marking capability

* Adding accept on ingress and egress marks

* Correcting an ingress marker

* Correcting unit test cases to show the appropriate markers

* Correcting a comment

* Addressing comments