Added the Pod Net ARM ID as a Label. (#1260)
* Added the Pod Net ARM ID as a Label.

Added Pod Net Arm Id as labels to the CNS Metrics.

* CNS Metrics: ARM ID as Label

Necessary changes to the read the ARM ID metadata coming from DNC.

* Subtle changes in NNC

Variable names had to be modified to meet lint standards.

* Subtle changes in NNC

Variable names had to be modified to meet lint standards.

* Made NNC arg a Pointer

NNC that was passed to create the ARM ID was too heavy. Converted that argument to a pointer and passed the reference of the NNC in that method.

* [CNS] Metrics Prefix

Added the "cx_" prefix to prometheus metrics of CNS.

* Minor Changes

Minor edits with the Metric Names and Documentation.

* ConstLabels in Metrics

Made changes to the ConstLabels in the CNS Metrics.

* ConstLabels in Metrics

Made changes to the ConstLabels in the CNS Metrics.