Add two more CNS endpoints to the CNS client (#1541)
* Add NumOfCPUCores method to CNS client

This is the last part of CNS's API surface that is used by DNC, but not
covered by a method from the CNS client.

* Remove unnecessary error wrapping lambda

The error handling here is simple enough that it doesn't really warrant
wrapping it with a lambda.

* Add NmAgentSupportedAPIs to the CNS Client

DNC uses this API to detect GRE key capabilities. Since it's not
supported by the client, it does this with direct HTTP requests
currently. In order to ensure that there's only one way of accessing
CNS, this implements the method so that the client can be used in DNC
instead.

* Add NMA supported APIs to client URL list

This was forgotten in a previous iteration.

* Handle non-200 status code by returning an error

The NMAgentSupportedAPIs method did not error when a non-2XX status code
was encountered. This leads to surprising behavior on the part of the
consumer.

* Use http.NoBody instead of nil

This was a linter suggestion, and a good one. http.NoBody is more
semantically meaningful than passing a nil.

* Create a FailedHTTPRequest error type

This is in response to a linter complaint that dynamic errors were being
used instead of a static error. Declaring an error type with var
wouldn't carry along necessary debugging information (namely the HTTP
status code), so it wasn't really appropriate. Rather than disable the
check entirely with a linter directive, this defines a reusable error
type so that HTTP errors can be communicated upward consistently.

* Use an alias for CNS client paths

These really should be a single set of paths, but there's this existing
block of constants to define a "contract with DNC." Whether or not
that's complete, the spirit of it is somewhat clear. However, it's not
great to be duplicating constants all over the place. This is somewhat
of a compromise by defining the newer constants in terms of the older
ones.

* Add missing context to outbound HTTP requests

This was a good suggestion from the linter.